Split the project dependencies into rule
